sr(api): move find-paginated-user-recommended-trainings in devcomp
